% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_template_tables}
\alias{db_template_tables}
\title{Template tables for the Peatland Decomposition Database}
\format{
A list where each element is a table in the Peatland Decomposition
Database. This is primarily for internal use.
}
\usage{
db_template_tables
}
\description{
Template tables for the Peatland Decomposition Database
}
\keyword{datasets}
