% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_get_dm.R
\name{dp_get_dm}
\alias{dp_get_dm}
\title{Access the dpeatdecomposition database as a \code{dm} object.}
\usage{
dp_get_dm(con, learn_keys = FALSE)
}
\arguments{
\item{con}{A connection to the dpeatdecomposition database. This should be an object of
class \code{\link[RMariaDB:MariaDBConnection]{MariaDBConnection}}, created
using \code{\link[RMariaDB:dbConnect]{dbConnect}}.}

\item{learn_keys}{A logical value indicating if primary and foreign keys
should be assigned (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
A \code{dm} object with the dpeatdecomposition database and all primary and
foreign keys.
}
\description{
\code{dp_get_dm} extracts tables and relations in the form of a
\code{\link[dm:dm]{dm}} object. It is a wrapper around
\code{\link[dm:dm_from_src]{dm_from_src}} which manually adds primary and
foreign keys to the tables which dm currently does not support for MariaDB
databases.
}
\examples{
\dontrun{
# connect to database
con <-
  RMariaDB::dbConnect( # ---todo: change this
    drv = RMariaDB::MariaDB(),
    dbname = "dpeatdecomposition",
    username = "root",
    password = "coucou",
    host = "mariadb"
  )

# get dpeatdecomposition database as dm object
dm_dpeatdecomposition <- dp_get_dm(con, learn_keys = TRUE)

# when finished, disconnect
RMariaDB::dbDisconnect(con)
}
}
