% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpeatdecomposition-package.R
\docType{package}
\name{dpeatdecomposition-package}
\alias{dpeatdecomposition}
\alias{dpeatdecomposition-package}
\title{dpeatdecomposition: R Interface to the Peatland Decomposition Database}
\description{
R Interface to the Peatland Decomposition Database. Provides functions to access the database using the 'dm' and 'RMariaDB' packages.
}
\author{
\strong{Maintainer}: Henning Teickner \email{henning.teickner@uni-muenster.de} (\href{https://orcid.org/0000-0002-3993-1182}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Edzer Pebesma \email{edzer.pebesma@uni-muenster.de} (\href{https://orcid.org/0000-0001-8049-7069}{ORCID})
  \item Knorr Klaus-Holger \email{klaus-holger.knorr@uni-muenster.de} (\href{https://orcid.org/0000-0003-4175-0214}{ORCID})
}

}
\keyword{internal}
