% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_udunits_to_eml.R
\name{pd_udunits_to_eml}
\alias{pd_udunits_to_eml}
\title{Converts a column with udunits as character values to the respective EML units.}
\usage{
pd_udunits_to_eml(x)
}
\arguments{
\item{x}{A character vector with values representing udunits units.}
}
\value{
A character vector with the same length as rows in \code{x} with
units according to the Ecological Metadata Language.
}
\description{
\code{pd_udunits_to_eml} takes a character vector with values representing
udunits units and converts these to units as used by the Ecological Metadata
Language. Note that the function converts only a specific sets of units as
needed for the Peatland Decomposition Database. All other units are converted
to \code{NA}.
}
\examples{
\dontrun{
# convert a known unit
pd_udunits_to_eml("mm")

# convert a unit not yet implemented
pd_udunits_to_eml("L")
}
}
\keyword{Internal}
