% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_get_dm.R
\name{dp_get_dm}
\alias{dp_get_dm}
\title{Access the Peatland Decomposition Database as a \code{dm} object.}
\usage{
dp_get_dm(con, learn_keys = FALSE)
}
\arguments{
\item{con}{A connection to the Peatland Decomposition Database. This should
be an object of class \code{MariaDBConnection}, created with \href{RMariaDB::dbConnect_MariaDBDriver}{RMariaDB::dbConnect()}.}

\item{learn_keys}{A logical value indicating if primary and foreign keys
should be assigned (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
A \code{dm} object with the Peatland Decomposition Database and all primary and
foreign keys.
}
\description{
\code{dp_get_dm} extracts tables and relations in the form of a
\code{dm} object. It is a wrapper around
\code{\link[dm:dm_from_src]{dm::dm_from_src()}} which manually adds primary and
foreign keys to the tables which dm currently does not support for MariaDB
databases.
}
\examples{
\dontrun{
# connect to database
con <-
  RMariaDB::dbConnect(
    drv = RMariaDB::MariaDB(),
    dbname = "dpeatdecomposition",
    default.file = "~/my.cnf"
  )

# get Peatland Decomposition Database as dm object
dm_dpeatdecomposition <- dp_get_dm(con, learn_keys = TRUE)

# disconnect
RMariaDB::dbDisconnect(con)
}
}
